refactor : sse package 구조
