+++ src/main/java/com/codeit/otboo/domain/sse/event/listener/SseHandler.java
@@ -1,4 +1,4 @@
-package com.codeit.otboo.domain.sse.util;
+package com.codeit.otboo.domain.sse.event.listener;
 
 import java.util.UUID;
 
@@ -8,6 +8,7 @@ import org.springframework.transaction.event.TransactionalEventListener;
 
 import com.codeit.otboo.domain.notification.dto.NotificationDto;
 import com.codeit.otboo.domain.sse.service.SseEmitterService;
+import com.codeit.otboo.domain.sse.util.SseMessage;
 import com.codeit.otboo.exception.CustomException;
 import com.codeit.otboo.global.error.ErrorCode;
